Extract all rdata types to rdata module
This is a huge refactor with a few breaking changes. The idea here
is to make it easier to add more resource types as well as making
it easier to maintain existing types.

All the logic for parsing RDATA has been moved to the rdata module.
The module is further broken down into individual resource types.
